memberships datacube documentation

memberships datacube documentation

memberships

Format

The memberships datacube is a list that contains the following 4 datasets: ECOLEX_MEM, IEADB_MEM, TFDD_MEM, HUGGO_MEM. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

ECOLEX_MEM:

A dataset with 24975 observations and the following 10 variables: manyID, treatyID, Title, Begin, stateID, End, stateSignature, Force, Rat, ecolexID.

IEADB_MEM:

A dataset with 15463 observations and the following 12 variables: manyID, treatyID, Title, Begin, stateID, End, stateSignature, Signature, Rat, Force, DocType, ieadbID.

TFDD_MEM:

A dataset with 1832 observations and the following 8 variables: manyID, treatyID, Title, Begin, stateID, Signature, tfddID, Memberships.

HUGGO_MEM:

A dataset with 50979 observations and the following 34 variables:manyID, Title, Begin, stateID, StateSignature, StateRatification, StateForce, StateEnd, Accession, Succession, treatyID, Signature, Force, End, gengID, ieaID, ecolexID, Coder, StateForce2, StateForce3, StateEnd2, Term, Comments, Deposit, obsolete, ProvisionalApp, Reservation, Notes, stateForce_ecolex, stateForce_iea, Consent, Acceptance, Orig_noneng_title match.

Source

ECOLEX_MEM: J. Sommer. “Global governance in forestry: a cross-national analysis”. International Journal of Sustainable Development & World Ecology 27.6 (2020), pp. 481-495.

IEADB_MEM: R. B. Mitchell, L. B. Andonova, M. Axelrod, et al. “What we know (and could know) about internationalenvironmental agreements”. Global Environmental Politics 20.1 (2020), pp. 103-121.

TFDD_MEM: O. College of Earth and O. S. U. Atmospheric Science. Product of the Transboundary Freshwater DisputeDatabase. http://transboundarywaters.science.oregonstate.edu.. 2021.

HUGGO_MEM: J. Hollway. Environmental agreements for manydata. 2021.

Details

#> $ECOLEX_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           24975|     0     |       0       |
#> |   treatyID    |  character  |           24975|     0     |       0       |
#> |     Title     |  character  |           24975|     0     |       0       |
#> |     Begin     |    mdate    |           24975|   3421    |     13.7      |
#> |    stateID    |  character  |           24975|    21     |     0.08      |
#> |      End      |    mdate    |           24975|   24511   |     98.14     |
#> |stateSignature |    mdate    |           24975|   15313   |     61.31     |
#> |     Force     |    mdate    |           24975|   5783    |     23.16     |
#> |      Rat      |    mdate    |           24975|   6993    |      28       |
#> |   ecolexID    |  character  |           24975|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IEADB_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           15463|     0     |       0       |
#> |   treatyID    |  character  |           15463|     0     |       0       |
#> |     Title     |  character  |           15463|     0     |       0       |
#> |     Begin     |    mdate    |           15463|     0     |       0       |
#> |    stateID    |  character  |           15463|     0     |       0       |
#> |      End      |    mdate    |           15463|   14593   |     94.37     |
#> |stateSignature |    mdate    |           15463|   5797    |     37.49     |
#> |   Signature   |    mdate    |           15463|     0     |       0       |
#> |      Rat      |    mdate    |           15463|   5485    |     35.47     |
#> |     Force     |    mdate    |           15463|     0     |       0       |
#> |    DocType    |  character  |           15463|     0     |       0       |
#> |    ieadbID    |   numeric   |           15463|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TFDD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1832|     0     |       0       |
#> |   treatyID    |  character  |            1832|     0     |       0       |
#> |     Title     |  character  |            1832|     1     |     0.05      |
#> |     Begin     |    mdate    |            1832|    102    |     5.57      |
#> |    stateID    |  character  |            1832|     1     |     0.05      |
#> |   Signature   |    mdate    |            1832|    102    |     5.57      |
#> |    tfddID     |   numeric   |            1832|     1     |     0.05      |
#> |  Memberships  |  character  |            1832|    805    |     43.94     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           50979|     0     |       0       |
#> |      Title      |  character  |           50979|     0     |       0       |
#> |      Begin      |    mdate    |           50979|    12     |     0.02      |
#> |     stateID     |  character  |           50979|     0     |       0       |
#> | StateSignature  |    mdate    |           50979|   11901   |     23.34     |
#> |StateRatification|    mdate    |           50979|   27691   |     54.32     |
#> |   StateForce    |    mdate    |           50979|   18858   |     36.99     |
#> |    StateEnd     |    mdate    |           50979|   32861   |     64.46     |
#> |    Accession    |    mdate    |           50979|   42973   |     84.3      |
#> |   Succession    |  character  |           50979|     0     |       0       |
#> |    treatyID     |  character  |           50979|     0     |       0       |
#> |    Signature    |    mdate    |           50979|    14     |     0.03      |
#> |      Force      |    mdate    |           50979|   3526    |     6.92      |
#> |       End       |    mdate    |           50979|   38634   |     75.78     |
#> |     gengID      |  character  |           50979|     0     |       0       |
#> |      ieaID      |  character  |           50979|   8626    |     16.92     |
#> |    ecolexID     |  character  |           50979|   30960   |     60.73     |
#> |      Coder      |  character  |           50979|     0     |       0       |
#> |   StateForce2   |    mdate    |           50979|   44274   |     86.85     |
#> |   StateForce3   |    mdate    |           50979|   47289   |     92.76     |
#> |    StateEnd2    |    mdate    |           50979|   47283   |     92.75     |
#> |      Term       |    mdate    |           50979|   49801   |     97.69     |
#> |    Comments     |  character  |           50979|   47699   |     93.57     |
#> |     Deposit     |  character  |           50979|   49329   |     96.76     |
#> |    obsolete     |  character  |           50979|   38384   |     75.29     |
#> | ProvisionalApp  |  character  |           50979|   50207   |     98.49     |
#> |   Reservation   |  character  |           50979|   50205   |     98.48     |
#> |      Notes      |  character  |           50979|   50793   |     99.64     |
#> |stateForce_ecolex|  character  |           50979|   32462   |     63.68     |
#> | stateForce_iea  |  character  |           50979|   32462   |     63.68     |
#> |     Consent     |  character  |           50979|   50755   |     99.56     |
#> |   Acceptance    |  character  |           50979|   50454   |     98.97     |
#> |Orig_noneng_title|  character  |           50979|   50975   |     99.99     |
#> |      match      |  character  |           50979|   50975   |     99.99     |
#> ------------------------------------------------------------------------------

URL

ECOLEX_MEM: https://www.ecolex.org/result/?type=treaty

IEADB_MEM: https://iea.uoregon.edu/country-members

TFDD_MEM: https://transboundarywaters.science.oregonstate.edu/

HUGGO_MEM: Hand-coded data by the GGO team.

Mapping

ECOLEX_MEM: Variable Mapping

from	to

For	Force

Rati	Rat

Sign	stateSignature

Term	End

StatID	stateID

EcolexID	ecolexID

IEADB_MEM: Variable Mapping

from	to

country	stateID

treatyname	Title

tsig	Signature

csig	stateSignature

crat	Rat

tterm	End

ceif3	Force

ceif4	Force2

mitch_id	ieadbID

inclusion	DocType

TFDD_MEM: Variable Mapping

from	to

DateSigned	Signature

'2016Update ID'	tfddID

CCODE	stateID

DocumentName	Title

Signatories	Memberships

HUGGO_MEM: Variable Mapping

from	to

NA	NA