references datacube documentation

references datacube documentation

references

Format

The references datacube is a list that contains the following 1 datasets: ECOLEX_REF. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

ECOLEX_REF:

A dataset with 1164 observations and the following 3 variables: treatyID1, treatyID2, RefType.

Source

ECOLEX_REF: R. E. Kim and B. Mackey. “International environmental law as a complex adaptive system”. In: International Environmental Agreements: Politics, Law and Economics 14.1 (2014), pp. 5-24.

Details

#> $ECOLEX_REF
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |   treatyID1   |  character  |            1164|     0     |       0       |
#> |   treatyID2   |  character  |            1164|     0     |       0       |
#> |    RefType    |  character  |            1164|     0     |       0       |
#> ----------------------------------------------------------------------------

URL

ECOLEX_REF: NA

Mapping

ECOLEX_REF: Variable Mapping

from	to

NA	NA