agreements database documentation

agreements database documentation

agreements

Format

The agreements database is a list that contains the following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

IEADB:

A dataset with 3666 observations and the following 10 variables: manyID, Title, Beg, DocType, AgreementType, Signature, Force, Lineage, treatyID, ieadbID.

ECOLEX:

A dataset with 2174 observations and the following 10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force, Lineage, treatyID, ecolexID.

CIESIN:

A dataset with 666 observations and the following 7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.

HEIDI:

A dataset with 2280 observations and the following 7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.

HUGGO:

A dataset with 7329 observations and the following 49 variables: manyID, treatyID, Title, Beg, End, Signature, Force, DocType, AgreementType, GeogArea, gnevarID, verified, gengID, ecolexID, ieaID, DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex, subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices, Depository, DepositoryURL, Published, Abstract, Website1, Website2, Secretariat, SecretariatURL, UNEP, Supersedes, References, EnabledBy, AmendedBy, Lit, Data, Coded, TreatyText, SourceText, Text, AbstractText.

AIGGO:

A dataset with 7294 observations and the following 13 variables: manyID, treatyID, Title, Beg, End, Signature, Force, action_area, linkage, accessionC, accessionP, termination_type, termination_date.

Details

#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3666|     0     |       0       |
#> |     Title     |  character  |            3666|     0     |       0       |
#> |      Beg      |    mdate    |            3666|    65     |     1.77      |
#> |    DocType    |  character  |            3666|     0     |       0       |
#> | AgreementType |  character  |            3666|     0     |       0       |
#> |   Signature   |    mdate    |            3666|    65     |     1.77      |
#> |     Force     |    mdate    |            3666|    867    |     23.65     |
#> |    Lineage    |  character  |            3666|     0     |       0       |
#> |   treatyID    |  character  |            3666|     0     |       0       |
#> |    ieadbID    |  character  |            3666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |      Beg      |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             666|     0     |       0       |
#> |     Title     |  character  |             666|     0     |       0       |
#> |      Beg      |    mdate    |             666|     0     |       0       |
#> |   Signature   |    mdate    |             666|     0     |       0       |
#> |     Force     |    mdate    |             666|     0     |       0       |
#> |    Lineage    |  character  |             666|     0     |       0       |
#> |   treatyID    |  character  |             666|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |      Beg      |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            7329|   3479    |     47.47     |
#> |    treatyID     |  character  |            7329|   3479    |     47.47     |
#> |      Title      |  character  |            7329|     0     |       0       |
#> |       Beg       |    mdate    |            7329|    60     |     0.82      |
#> |       End       |    mdate    |            7329|   6653    |     90.78     |
#> |    Signature    |    mdate    |            7329|   3536    |     48.25     |
#> |      Force      |    mdate    |            7329|   5168    |     70.51     |
#> |     DocType     |  character  |            7329|   4565    |     62.29     |
#> |  AgreementType  |  character  |            7329|   4603    |     62.81     |
#> |    GeogArea     |  character  |            7329|   5185    |     70.75     |
#> |    gnevarID     |  character  |            7329|   4569    |     62.34     |
#> |    verified     |  character  |            7329|   7280    |     99.33     |
#> |     gengID      |  character  |            7329|   6410    |     87.46     |
#> |    ecolexID     |  character  |            7329|   6842    |     93.36     |
#> |      ieaID      |  character  |            7329|   6619    |     90.31     |
#> |DocValidUntilDate|  character  |            7329|   7295    |     99.54     |
#> |      Notes      |  character  |            7329|   7281    |     99.35     |
#> |    Download     |  character  |            7329|   6410    |     87.46     |
#> |    MEA_type     |  character  |            7329|   6619    |     90.31     |
#> |      Ambit      |  character  |            7329|   6900    |     94.15     |
#> |     Region      |  character  |            7329|   7079    |     96.59     |
#> | subject_ecolex  |  character  |            7329|   6842    |     93.36     |
#> |   subject_iea   |  character  |            7329|   6619    |     90.31     |
#> |    Keywords     |  character  |            7329|   6914    |     94.34     |
#> |     Lineage     |  character  |            7329|   6619    |     90.31     |
#> |    Sequence     |  character  |            7329|   6620    |     90.33     |
#> |    AdoptedIn    |  character  |            7329|   6895    |     94.08     |
#> |    Languages    |  character  |            7329|   6842    |     93.36     |
#> |   Appendices    |  character  |            7329|   7294    |     99.52     |
#> |   Depository    |  character  |            7329|   6936    |     94.64     |
#> |  DepositoryURL  |  character  |            7329|   7312    |     99.77     |
#> |    Published    |  character  |            7329|   7061    |     96.34     |
#> |    Abstract     |  character  |            7329|   7302    |     99.63     |
#> |    Website1     |  character  |            7329|   6942    |     94.72     |
#> |    Website2     |  character  |            7329|   7257    |     99.02     |
#> |   Secretariat   |  character  |            7329|   6620    |     90.33     |
#> | SecretariatURL  |  character  |            7329|   6620    |     90.33     |
#> |      UNEP       |  character  |            7329|   7127    |     97.24     |
#> |   Supersedes    |  character  |            7329|   7312    |     99.77     |
#> |   References    |  character  |            7329|   7293    |     99.51     |
#> |    EnabledBy    |  character  |            7329|   7220    |     98.51     |
#> |    AmendedBy    |  character  |            7329|   7238    |     98.76     |
#> |       Lit       |  character  |            7329|   6775    |     92.44     |
#> |      Data       |  character  |            7329|   6783    |     92.55     |
#> |      Coded      |  character  |            7329|   7024    |     95.84     |
#> |   TreatyText    |  character  |            7329|   5776    |     78.81     |
#> |   SourceText    |  character  |            7329|   5776    |     78.81     |
#> |      Text       |  character  |            7329|   7048    |     96.17     |
#> |  AbstractText   |  character  |            7329|   7048    |     96.17     |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            7294|   3479    |     47.7      |
#> |    treatyID    |  character  |            7294|   3479    |     47.7      |
#> |     Title      |  character  |            7294|     0     |       0       |
#> |      Beg       |    mdate    |            7294|    56     |     0.77      |
#> |      End       |    mdate    |            7294|   6621    |     90.77     |
#> |   Signature    |    mdate    |            7294|   3532    |     48.42     |
#> |     Force      |    mdate    |            7294|   5150    |     70.61     |
#> |  action_area   |  character  |            7294|    769    |     10.54     |
#> |    linkage     |  character  |            7294|   4782    |     65.56     |
#> |   accessionC   |  character  |            7294|   1268    |     17.38     |
#> |   accessionP   |  character  |            7294|   7032    |     96.41     |
#> |termination_type|  character  |            7294|   6765    |     92.75     |
#> |termination_date|  character  |            7294|   7074    |     96.98     |
#> -----------------------------------------------------------------------------