agreements datacube documentation

agreements datacube documentation

agreements

Format

The agreements datacube is a list that contains the following 7 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO, IRD. For more information and references to each of the datasets used, please use the manydata::call_sources() and manydata::compare_dimensions() functions.

IEADB:

A dataset with 3667 observations and the following 10 variables: manyID, Title, Begin, DocType, AgreementType, Signature, Force, Lineage, treatyID, ieadbID.

ECOLEX:

A dataset with 2174 observations and the following 10 variables: manyID, Title, Begin, DocType, GeogArea, Signature, Force, Lineage, treatyID, ecolexID.

CIESIN:

A dataset with 666 observations and the following 7 variables: manyID, Title, Begin, Signature, Force, Lineage, treatyID.

HEIDI:

A dataset with 2280 observations and the following 7 variables: manyID, Title, Begin, Signature, Lineage, treatyID, heidiID.

HUGGO:

A dataset with 3982 observations and the following 51 variables: manyID, treatyID, Title, Begin, End, Signature, Force, url, AgreementType, DocType, GeogArea, gengID, ieaID, ecolexID, Parties, verified, DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex, subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices, Depository, DepositoryURL, Published, Website1, Website2, Secretariat, SecretariatURL, UNEP, Supersedes, References, EnabledBy, AmendedBy, Lit, Data, Coded, Abstract, TreatyText, Language, Orig_noneng_title, match, Coder.

AIGGO:

A dataset with 3923 observations and the following 13 variables: manyID, treatyID, Title, Begin, End, Signature, Force, action_area, linkage, accessionC, accessionP, termination_type, termination_date.

IRD:

A dataset with 92 observations and the following 9 variables: manyID, treatyID, Title, Begin, RegimeComponent, Wat, Wat2, End, RegimeElement.

Source

IEADB: R. B. Mitchell et al. "What we know (and could know) about international environmental agreements". Global Environmental Politics 20.1 (2020), pp. 103-121.

ECOLEX: J. Sommer. “Global governance in forestry: a cross-national analysis”. International Journal ofSustainable Development & World Ecology 27.6 (2020), pp. 481-495.

CIESIN: Centre for International Earth Science Information Network. Socioeconomic Data and Applications Center (SEDAC) Collection of Treaty Texts (2002).

HEIDI: J-F. Morin, and C. Blouin. "How environmental treaties contribute to global health governance". Globalization and health 15.1 (2019), pp. 1-8.

HUGGO: J. Hollway. Environmental agreements for manydata. 2021.

AIGGO: J. Hollway. Environmental agreements for manydata. 2021.

IRD: O. R. Young and M. Zürn. “The international regimes database: Designing and using a sophisticated tool for institutional analysis”. In: Global Environmental Politics 6.3 (2006), pp. 121-143.

Details

#> $IEADB
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            3667|     0     |       0       |
#> |   treatyID    |  character  |            3667|     0     |       0       |
#> |     Title     |  character  |            3667|     0     |       0       |
#> |     Begin     |    mdate    |            3667|     0     |       0       |
#> |    DocType    |  character  |            3667|     0     |       0       |
#> | AgreementType |  character  |            3667|     0     |       0       |
#> |   Signature   |    mdate    |            3667|    65     |     1.77      |
#> |     Force     |    mdate    |            3667|    868    |     23.67     |
#> |    Lineage    |  character  |            3667|     0     |       0       |
#> |    ieadbID    |  character  |            3667|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ECOLEX
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2174|     0     |       0       |
#> |   treatyID    |  character  |            2174|     0     |       0       |
#> |     Title     |  character  |            2174|     0     |       0       |
#> |     Begin     |    mdate    |            2174|     0     |       0       |
#> |    DocType    |  character  |            2174|     0     |       0       |
#> |   GeogArea    |  character  |            2174|    90     |     4.14      |
#> |   Signature   |    mdate    |            2174|     0     |       0       |
#> |     Force     |    mdate    |            2174|   1066    |     49.03     |
#> |    Lineage    |  character  |            2174|     0     |       0       |
#> |   ecolexID    |  character  |            2174|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $CIESIN
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             662|     0     |       0       |
#> |     Title     |  character  |             662|     0     |       0       |
#> |     Begin     |    mdate    |             662|     0     |       0       |
#> |   Signature   |    mdate    |             662|     0     |       0       |
#> |     Force     |    mdate    |             662|    182    |     27.49     |
#> |    Lineage    |  character  |             662|    208    |     31.42     |
#> |   treatyID    |  character  |             662|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HEIDI
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2280|     0     |       0       |
#> |   treatyID    |  character  |            2280|     0     |       0       |
#> |     Title     |  character  |            2280|     0     |       0       |
#> |     Begin     |    mdate    |            2280|    27     |     1.18      |
#> |   Signature   |    mdate    |            2280|    27     |     1.18      |
#> |    Lineage    |  character  |            2280|     0     |       0       |
#> |    heidiID    |   numeric   |            2280|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            3982|     0     |       0       |
#> |    treatyID     |  character  |            3982|     0     |       0       |
#> |      Title      |  character  |            3982|     0     |       0       |
#> |      Begin      |    mdate    |            3982|    57     |     1.43      |
#> |       End       |    mdate    |            3982|   3047    |     76.52     |
#> |    Signature    |    mdate    |            3982|    57     |     1.43      |
#> |      Force      |    mdate    |            3982|   1587    |     39.85     |
#> |       url       |  character  |            3982|   2688    |     67.5      |
#> |  AgreementType  |  character  |            3982|   1279    |     32.12     |
#> |     DocType     |  character  |            3982|   1243    |     31.22     |
#> |    GeogArea     |  character  |            3982|   1842    |     46.26     |
#> |     gengID      |  character  |            3982|    197    |     4.95      |
#> |      ieaID      |  character  |            3982|   3286    |     82.52     |
#> |    ecolexID     |  character  |            3982|   3504    |      88       |
#> |     Parties     |  character  |            3982|   1711    |     42.97     |
#> |    verified     |  character  |            3982|   3934    |     98.79     |
#> |DocValidUntilDate|  character  |            3982|   3948    |     99.15     |
#> |      Notes      |  character  |            3982|   3935    |     98.82     |
#> |    Download     |  character  |            3982|   3083    |     77.42     |
#> |    MEA_type     |  character  |            3982|   3286    |     82.52     |
#> |      Ambit      |  character  |            3982|   3563    |     89.48     |
#> |     Region      |  character  |            3982|   3735    |     93.8      |
#> | subject_ecolex  |  character  |            3982|   3505    |     88.02     |
#> |   subject_iea   |  character  |            3982|   3286    |     82.52     |
#> |    Keywords     |  character  |            3982|   3575    |     89.78     |
#> |     Lineage     |  character  |            3982|   3286    |     82.52     |
#> |    Sequence     |  character  |            3982|   3287    |     82.55     |
#> |    AdoptedIn    |  character  |            3982|   3555    |     89.28     |
#> |    Languages    |  character  |            3982|   3505    |     88.02     |
#> |   Appendices    |  character  |            3982|   3947    |     99.12     |
#> |   Depository    |  character  |            3982|   3595    |     90.28     |
#> |  DepositoryURL  |  character  |            3982|   3966    |     99.6      |
#> |    Published    |  character  |            3982|   3721    |     93.45     |
#> |    Website1     |  character  |            3982|   3600    |     90.41     |
#> |    Website2     |  character  |            3982|   3958    |     99.4      |
#> |   Secretariat   |  character  |            3982|   3287    |     82.55     |
#> | SecretariatURL  |  character  |            3982|   3287    |     82.55     |
#> |      UNEP       |  character  |            3982|   3780    |     94.93     |
#> |   Supersedes    |  character  |            3982|   3965    |     99.57     |
#> |   References    |  character  |            3982|   3947    |     99.12     |
#> |    EnabledBy    |  character  |            3982|   3875    |     97.31     |
#> |    AmendedBy    |  character  |            3982|   3891    |     97.71     |
#> |       Lit       |  character  |            3982|   3437    |     86.31     |
#> |      Data       |  character  |            3982|   3444    |     86.49     |
#> |      Coded      |  character  |            3982|   3683    |     92.49     |
#> |    Abstract     |  character  |            3982|   3956    |     99.35     |
#> |   TreatyText    |  character  |            3982|   2632    |     66.1      |
#> |    Language     |  character  |            3982|   2632    |     66.1      |
#> |Orig_noneng_title|  character  |            3982|   3926    |     98.59     |
#> |      match      |  character  |            3982|   3926    |     98.59     |
#> |      Coder      |  character  |            3982|     0     |       0       |
#> ------------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |            3923|     0     |       0       |
#> |    treatyID    |  character  |            3923|     0     |       0       |
#> |     Title      |  character  |            3923|     0     |       0       |
#> |     Begin      |    mdate    |            3923|    47     |      1.2      |
#> |      End       |    mdate    |            3923|   2998    |     76.42     |
#> |   Signature    |    mdate    |            3923|    47     |      1.2      |
#> |     Force      |    mdate    |            3923|   1548    |     39.46     |
#> |  action_area   |  character  |            3923|    452    |     11.52     |
#> |    linkage     |  character  |            3923|   3283    |     83.69     |
#> |   accessionC   |  character  |            3923|    708    |     18.05     |
#> |   accessionP   |  character  |            3923|   3923    |      100      |
#> |termination_type|  character  |            3923|   3828    |     97.58     |
#> |termination_date|  character  |            3923|   3828    |     97.58     |
#> -----------------------------------------------------------------------------
#> 
#> 
#> $IRD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |              92|     0     |       0       |
#> |   treatyID    |  character  |              92|     0     |       0       |
#> |     Title     |  character  |              92|     0     |       0       |
#> |     Begin     |    mdate    |              92|    36     |     39.13     |
#> |RegimeComponent|  character  |              92|    33     |     35.87     |
#> |      Wat      |    mdate    |              92|    34     |     36.96     |
#> |     Wat2      |    mdate    |              92|    78     |     84.78     |
#> |      End      |    mdate    |              92|    36     |     39.13     |
#> | RegimeElement |  character  |              92|     0     |       0       |
#> ----------------------------------------------------------------------------

URL

IEADB: https://iea.uoregon.edu/base-agreement-list

ECOLEX: https://www.ecolex.org/result/?type=treaty

CIESIN: https://sedac.ciesin.columbia.edu/entri/

HEIDI: https://www.chaire-epi.ulaval.ca/en/data/heidi

HUGGO: Hand-coded data by the GGO team.

AIGGO: Machine-generated data by the GGO team.

IRD: https://iea.uoregon.edu/ird

Mapping

IEADB: Variable Mapping

from	to

'IEA# (click for add'l info)'	ieadbID

'Treaty Name'	Title

'Signature Date'	Signature

'Date IEA entered into force'	Force

'Agreement Type (level 2)'	AgreementType

Inclusion	DocType

ECOLEX: Variable Mapping

from	to

EcolexID	ecolexID

title	Title

Date	Signature

'Entry.into.Force'	Force

Document.Type	DocType

Field.of.application	GeogArea

CIESIN: Variable Mapping

from	to

'Treaty Title'	Title

'Year of Agreement'	Signature

'Year of Entry into Force'	Force

HEIDI: Variable Mapping

from	to

ID	heidiID

'Name.of.the.agreement'	Title

signature.date	Signature

IRD: Variable Mapping

from	to

Regime	Title

Formation	Begin

Endpoint	End

Watershed	Wat

Watershed2	Wat2

HUGGO: Variable Mapping

from	to

NA	NA

AIGGO: Variable Mapping

from	to

NA	NA